requirements.txt
aioredis==1.3.1
amqp==5.0.5
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-split-settings==1.0.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kombu==5.0.2
msgpack==1.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
prompt-toolkit==3.0.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.2.0
